requirements.txt
torch
torchvision
torchaudio
transformers
sentence-transformers
Pillow
flask
opencv-python
ffmpeg-python
tqdm
scikit-learn
numpy
